+++ MagicAPI/src/main/java/com/elmakers/mine/bukkit/api/magic/MageController.java
@@ -372,6 +372,7 @@ public interface MageController {
     String getBlockSkin(Material blockType);
     String getMobSkin(EntityType mobType);
     void checkResourcePack(CommandSender sender);
+    boolean sendResourcePack(CommandSender sender);
     boolean commitOnQuit();
     
     Set<String> getItemKeys();